+++ tests/download_wis_jma_to_cache.sh
@@ -40,18 +40,19 @@ fi
 err_num=`grep -F '[ERROR]' wis_jma_log.txt | wc -l`
 if test ${err_num} -eq 0; then
     grep "ETag:" wis_jma_log.txt | tail -1 | cut -d' ' -f2 > wis_jma_etag.txt
-    if test ${is_pre} -eq 1; then
+    if test ${is_pre} -eq 1 -a -s wis_jma_updated.txt; then
         updated_num=`cat wis_jma_updated.txt | wc -l`
         mkdir -p cache
         rm -rf wis_jma
         rm -f wis_jma_get2.txt
+	rm -f wis_jma_to_cache.txt
         while test ${updated_num} -ne 0; do
             rm -f wis_jma_log.txt
             mkdir -p wis_jma
 	    set +e
             aria2c -j ${parallel} -s ${parallel} -x ${parallel} --header 'Cache-Control: no-cache' --auto-file-renaming=false --allow-overwrite=false --log-level=error -l wis_jma_log.txt -i wis_jma_updated.txt -d wis_jma >> wis_jma_get2.txt
 	    set -e
-            met_pre_batch_to_cache RJTD wis_jma cache
+            met_pre_batch_to_cache RJTD wis_jma cache 1>> wis_jma_to_cache.txt 2>> met_pre_batch_to_cache.log
             grep -F '[ERROR]' wis_jma_log.txt | grep 'URI=' | sed -e 's/^.*URI=//g' | grep -v '^ *$' | sort -u > wis_jma_updated.txt
             updated_num=`cat wis_jma_updated.txt | wc -l`
             rm -rf wis_jma
